- Improved Angular component to handle dynamic input types (`IInterest` and `IChooseLanguageWithLevel`) - Enhanced form validation: submit button activation based on form state - Handled item deletion from array and reset form after submissionelete languages -add form and logic to handle language and interest values in new-post-component

+++ langauge-journal/src/app/badge-button-from-selection/badge-button-from-selection.component.ts
@@ -11,6 +11,11 @@ import { EventEmitter } from '@angular/core';
 import { InputPostBindingService } from '../input-post-binding.service';
 import { IChooseLanguageWithLevel } from '../types/Language/languageOptionTypes';
 import { ApiService } from '../api-service.service';
+import { IInterest } from '../types/Response/getInterestsResponse';
+import { HelperService } from '../helper.service';
+import { InterestService } from '../interest.service';
+
+type arrayOfInterest = Array<IChooseLanguageWithLevel | IInterest>;
 
 @Component({
   selector: 'app-badge-button-from-selection',
@@ -20,15 +25,16 @@ import { ApiService } from '../api-service.service';
   styleUrl: './badge-button-from-selection.component.scss',
 })
 export class BadgeButtonFromSelectionComponent implements OnInit, OnChanges {
-  public interestName: string = '';
   public arrayInterests: string[] = [];
-  @Input() selectedInterest: IChooseLanguageWithLevel[];
+  @Input() selectedInterest: arrayOfInterest;
   @Output() notifyParent: EventEmitter<IChooseLanguageWithLevel[]> =
     new EventEmitter();
 
   constructor(
     private inputBindingsService: InputPostBindingService,
     private apiService: ApiService,
+    private helperService: HelperService,
+    private interestService: InterestService,
   ) {}
 
   ngOnInit() {}
@@ -36,7 +42,7 @@ export class BadgeButtonFromSelectionComponent implements OnInit, OnChanges {
   ngOnChanges(changes: SimpleChanges) {
     if (this.selectedInterest[0].type === 'language') {
       const ArrayOfKeys = this.selectedInterest.map(
-        (object: IChooseLanguageWithLevel) => object.name,
+        (object: IChooseLanguageWithLevel | IInterest) => object.name,
       );
       this.arrayInterests = ArrayOfKeys;
       this.inputBindingsService.updateValueInterests(ArrayOfKeys);
@@ -46,21 +52,39 @@ export class BadgeButtonFromSelectionComponent implements OnInit, OnChanges {
   removeBadge(event: Event) {
     event.preventDefault();
     const id = (event.currentTarget as HTMLButtonElement).id;
-    const removeIndex = this.selectedInterest
-      .map((item) => item.name)
-      .indexOf(id);
-    const itemToDelete = this.selectedInterest[removeIndex];
-    this.apiService.deleteLanguage(itemToDelete).subscribe({
-      next: (value) => {
-        console.log(value);
-      },
-      error: (error) => {
-        console.log(error);
-      },
-    });
-    removeIndex >= 0 && this.selectedInterest.splice(removeIndex, 1);
-    this.notifyParent.emit(this.selectedInterest);
+    // const removeIndex = this.selectedInterest
+    //   .map((item) => item.name)
+    //   .indexOf(id);
+    // const itemToDelete = this.selectedInterest[removeIndex];
+    const itemToDelete = this.helperService.removeItemFromArray(
+      this.selectedInterest,
+      id,
+    );
+    if (this.isLanguage(itemToDelete)) {
+      this.apiService.deleteLanguage(itemToDelete).subscribe({
+        next: (value) => {
+          console.log(value);
+        },
+        error: (error) => {
+          console.log(error);
+        },
+      });
+    }
+
+    if (this.isInterest(itemToDelete)) {
+      this.apiService.deleteUserInterest(itemToDelete.interest_id).subscribe({
+        next: (value) => {
+          this.interestService.deleteInterest(itemToDelete);
+        },
+      });
+    }
   }
 
-  protected readonly Object = Object;
+  private isLanguage(item: any): item is IChooseLanguageWithLevel {
+    return item && item.type === 'language';
+  }
+
+  private isInterest(item: any): item is IInterest {
+    return item && item.type === 'interest';
+  }
 }
